utils/weather_reader.py
import aiofiles
import pandas as pd
import logging
import io

class WeatherFileReader:
    def __init__(self, file_path: str):
        self._file_path = file_path
        self.logger = logging.getLogger(__name__)

    def read_file(self):
        '''
        Reads the weather data from the file.

        Returns:
        A pandas DataFrame containing the weather data if successful, otherwise None.
        '''
        try:
            self.logger.info(f"Attempting to read file: {self._file_path}")
            
            # Replace async with regular file reading
            with open(self._file_path, mode='r') as f:
                content = f.read()

            df = pd.read_csv(io.StringIO(content))
            self.logger.info(f"Successfully read file: {self._file_path}")
            return df
        except FileNotFoundError:
            self.logger.error(f"File not found: {self._file_path}")
            raise
        except pd.errors.EmptyDataError:
            self.logger.error(f"File is empty: {self._file_path}")
            raise
        except pd.errors.ParserError:
            self.logger.error(f"Unable to parse file: {self._file_path}")
            raise
        except Exception as e:
            self.logger.error(f"Unexpected error reading file {self._file_path}: {str(e)}")
            return None
